Resolved some issues in moreno
Moreno now uses a programmatically defined protocol which helps with stability. I have temporarily added plotting of the protocols and outputs of the summary statistics while it is being tested. RDUB is ready to be included in the protocol but needs defining in the runMoreno function. Current only the first 1ms is used for the majority of protocols (requiring 100 points). I should use a resolution of 200 points per ms across the whole window to match Moreno et al 2016