Began simulated annealing Vanier approach
This is the first part of implementing this approach. A standard nelder mead implementation has been made that will then be edited to match Press "Numerical Recipes in C" and then changed again to match the final simulated annealing algorithm. This nelder mead implementation has been checked against (and matches) the scipy implementation