each bm now has a rmse function
previously the cost function for moreno overrode the bm cost function in order to do a weighted rmse cost function. However, since the tracker update in bm doesn't account for this, the tracker saw different (not weighted) costs. Now each bm has a rmse function that defines this calculation and moreno overrides this instead of overriding the cost function