initial sensitivities set based on steady state
big commit so buckle in
we run all of the models from steady state but that means the initial states are functions of parameters and this should be reflected in the sensitivities and the gradient calculations

for the HH models, we use MyGrad (an automatic differentiation package which allows normal numpy options so no need to replace numpy). by making the parameters into mygrad tensors we can use an internal myokit steady state function to calculate the sensitivities of the states and pass these back to myokit. We can't just use the normal analyticalModel.steady_state() because it applies some input sanitation that breaks mygrad so we have to use the hidden _steady_state_function

this kind of thing wont work for the markov model problems that calculate their steady states in a different way. there is some input sanitation in analyticalModel.steady_state() that can be avoided by using the internal _matrix_function() but this forces the use of numpy arrays rather than mygrad tensors. to get around this we need to duplicate this code (_matrix_function is dynamically generated so cant just read its source code) which is now in a new module called utils.autodiff. from this we can get the transition matrices behaving nicely with mygrad but numpy.linalg.solve (and basically any other matrix equation solver) will drop this information. so we also need a matrix Ax=b solver that plays nicely with mygrad which I had to just do by myself (now in the utils.autodiff module).

Most of this is quite messy and hopefully doesn't cause future headaches. I have put in checks to make sure the steady states calculated agree with myokit. I'll also put in unit tests soon to make sure the steady states and the initial sensitivities both make sense and work properly