fixed bug in staircase
bug in staircase problems introduced by the larger sampling size. Would cause jumps over steps in the protocol. Normally this would be resolved by myokit but the staircase protocol needs to be a time series (because of the ramps) so that won't do it automatically. Setting a max step size to smaller than the smallest step width also fixes this