improved dokos curvilinear robustness
a big issue with this optimiser is that it requires a bounding interval for the brent optimisation with the center cost lower than the bounds. sometimes this is hard to find and so an error is thrown. we now check an extra point to try and find one, if one can't be found we assume the optimiser has converged (a step of 1e-9 is too big)