bug fix in grad cache
This is a bit of an awkward bug to explain.
The lru_cache we use for the cost and grad like hashable types but we tend to work with numpy arrays instead. For the inputs, the parameter vector is converted to a tuple. However, I didn't realise that the cache returns by reference. So if you get the grad back as a solve (one that was solved and is now stored in the cache) and you edit that variable, you also will edit future returns from the cache.
This should only happen when returnCost=True (which we have in a couple of places) and where caching is useful (sometimes the same parameters get evaluated multiple times). This happens because there is an implicit conversion to a tuple (which normally stops this as it requires copying) but does save the grad when returnCost=True as the returned value will be (cost, grad) where grad is still a numpy array, and therefore still passed by reference.
Luckily, these don't ever coincide for the current approaches, but we have a fix here to just force a deepcopy.